Add tests and benchmark for esimd_resadd_norm_gemv_int4_pert
- test_resadd_norm_gemv_int4.py: correctness vs fp16 ref, zero_hidden
  edge case, INT4 vs FP8 comparison with output values printed
- bench_resadd_norm_gemv_int4.py: latency comparison INT4 vs FP8
- Fix ambiguous operator+= in resadd_norm_gemv_int4.h (float cast)

+++ vllm/custom-esimd-kernels-vllm/csrc/xpu/esimd_kernels/resadd_norm_gemv_int4.h
@@ -66,7 +66,7 @@ struct ResAddNormGEMV_int4_pert_kernel {
             sq.select<8,1>(0)  += sq.select<8,1>(8);
             sq.select<4,1>(0)  += sq.select<4,1>(4);
             sq.select<2,1>(0)  += sq.select<2,1>(2);
-            sum_sq += sq[0] + sq[1];
+            sum_sq += (float)sq[0] + (float)sq[1];
         }
 
         float inv_rms = sycl::ext::intel::esimd::rsqrt(
